+++ src/components/Typography/typography.tsx
@@ -4,7 +4,13 @@ import type { ComponentProps, FC } from "react";
 import { CodeBlock } from "@/components/CodeBlock";
 import { useTheme } from "@/components/ThemeProvider";
 
-import styles from "./Typography.module.scss";
+import styles from "./typography.module.scss";
+import {
+  TypographyH1,
+  TypographyH2,
+  TypographyH3,
+  TypographyH4,
+} from "./typography-headings";
 
 // Typography.Root
 // =============================================================================
@@ -18,121 +24,6 @@ export const TypographyRoot: FC<TypographyRootProps> = (props) => {
     <div className={classNames(styles.typographyRoot, className)} {...rest}>
       {children}
     </div>
-  );
-};
-
-// Typography.H1
-// =============================================================================
-
-type TypographyH1Props = ComponentProps<"h1"> & {
-  mt?: string;
-  mb?: string;
-};
-
-const TypographyH1: FC<TypographyH1Props> = (props) => {
-  const { children, className, style, mt, mb, ...rest } = props;
-
-  const computedStyle = {
-    marginTop: mt,
-    marginBottom: mb,
-    ...style,
-  };
-
-  return (
-    <h1
-      className={classNames(styles.h1, className)}
-      style={computedStyle}
-      {...rest}
-    >
-      {children}
-    </h1>
-  );
-};
-
-// Typography.H2
-// =============================================================================
-
-type TypographyH2Props = ComponentProps<"h2"> & {
-  mt?: string;
-  mb?: string;
-};
-
-const TypographyH2: FC<TypographyH2Props> = (props) => {
-  const { children, className, style, mt, mb, ...rest } = props;
-
-  const theme = useTheme();
-
-  const computedStyle = {
-    "--typography-h2-border-color": theme.colors.gray["200"],
-    marginTop: mt,
-    marginBottom: mb,
-    ...style,
-  };
-
-  return (
-    <h2
-      className={classNames(styles.h2, className)}
-      style={computedStyle}
-      {...rest}
-    >
-      {children}
-    </h2>
-  );
-};
-
-// Typography.H3
-// =============================================================================
-
-type TypographyH3Props = ComponentProps<"h3"> & {
-  mt?: string;
-  mb?: string;
-};
-
-const TypographyH3: FC<TypographyH3Props> = (props) => {
-  const { children, className, style, mt, mb, ...rest } = props;
-
-  const computedStyle = {
-    marginTop: mt,
-    marginBottom: mb,
-    ...style,
-  };
-
-  return (
-    <h3
-      className={classNames(styles.h3, className)}
-      style={computedStyle}
-      {...rest}
-    >
-      {children}
-    </h3>
-  );
-};
-
-// Typography.H4
-// =============================================================================
-
-type TypographyH4Props = ComponentProps<"h4"> & {
-  mt?: string;
-  mb?: string;
-};
-
-const TypographyH4: FC<TypographyH4Props> = (props) => {
-  const { children, className, style, mt, mb, ...rest } = props;
-
-  const computedStyle = {
-    marginTop: mt,
-    marginBottom: mb,
-    ...style,
-  };
-
-  return (
-    <h4
-      className={classNames(styles.h4, className)}
-      style={computedStyle}
-      {...rest}
-    >
-      {children}
-    </h4>
   );
 };
 
@@ -263,7 +154,7 @@ export const TypographyCodeBlock: FC<TypographyCodeBlockProps> = (props) => {
 // Typography
 // =============================================================================
 
-type TypographyType = {
+type ComponentTypography = FC<TypographyRootProps> & {
   Root: typeof TypographyRoot;
   H1: typeof TypographyH1;
   H2: typeof TypographyH2;
@@ -276,17 +167,21 @@ type TypographyType = {
   CodeBlock: typeof TypographyCodeBlock;
 };
 
-const Typography: TypographyType = {
-  Root: TypographyRoot,
-  H1: TypographyH1,
-  H2: TypographyH2,
-  H3: TypographyH3,
-  H4: TypographyH4,
-  P: TypographyP,
-  Ul: TypographyUl,
-  Li: TypographyLi,
-  Code: TypographyCode,
-  CodeBlock: TypographyCodeBlock,
+const Typography: ComponentTypography = (props: TypographyRootProps) => {
+  return <TypographyRoot {...props} />;
 };
+
+Typography.Root = TypographyRoot;
+Typography.H1 = TypographyH1;
+Typography.H2 = TypographyH2;
+Typography.H3 = TypographyH3;
+Typography.H4 = TypographyH4;
+Typography.P = TypographyP;
+Typography.Ul = TypographyUl;
+Typography.Li = TypographyLi;
+Typography.Code = TypographyCode;
+Typography.CodeBlock = TypographyCodeBlock;
+
+Typography.displayName = "Typography";
 
 export { Typography };